Create tools routes
