+++ bin/fragments.js
@@ -1,9 +1,12 @@
-const fetch = require('node-fetch');
-const fs = require('fs');
+const fetch = require('node-fetch')
+const fs = require('fs')
 
 fetch(`${process.env.API_URL}/graphql`, {
   method: 'POST',
-  headers: { 'Content-Type': 'application/json', Cookie: 'JSESSIONID=YOUR_SESSION_ID' },
+  headers: {
+    'Content-Type': 'application/json',
+    Cookie: 'JSESSIONID=YOUR_SESSION_ID',
+  },
   body: JSON.stringify({
     variables: {},
     query: `
@@ -21,18 +24,22 @@ fetch(`${process.env.API_URL}/graphql`, {
     `,
   }),
 })
-  .then(result => result.json())
-  .then(result => {
+  .then((result) => result.json())
+  .then((result) => {
     // here we're filtering out any type information unrelated to unions or interfaces
     console.log('result: ', result)
     result.data.__schema.types = result.data.__schema.types.filter(
-      type => type.possibleTypes !== null,
-    );
-    fs.writeFile('./src/types/fragmentTypes.json', JSON.stringify(result.data), err => {
-      if (err) {
-        console.error('Error writing fragmentTypes file', err);
-      } else {
-        console.log('Fragment types successfully extracted!');
-      }
-    });
-  });
+      (type) => type.possibleTypes !== null,
+    )
+    fs.writeFile(
+      './src/types/fragmentTypes.json',
+      JSON.stringify(result.data),
+      (err) => {
+        if (err) {
+          console.error('Error writing fragmentTypes file', err)
+        } else {
+          console.log('Fragment types successfully extracted!')
+        }
+      },
+    )
+  })
