+++ src/features/tools/charges/index.tsx
@@ -112,7 +112,7 @@ const Row: React.FC<{
   </>
 )
 
-export const ChargePage: React.FC = () => {
+export const ChargesPage: React.FC = () => {
   const [confirming, setConfirming] = useState(false)
   const { data, loading, error } = usePaymentScheduleQueryQuery({
     variables: {
