+++ src/features/tools/staging-tools/unsign-member-tool/index.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { toast } from 'react-hot-toast'
 import { useUnsignMemberMutation } from 'types/generated/graphql'
 
-export const UnsignMemberTool: React.FC = () => {
+export const UnsignMemberPage: React.FC = () => {
   const [ssn, setSsn] = React.useState('')
   const [useUnsignMember, { loading }] = useUnsignMemberMutation()
